requirements.txt
python-dotenv
telegraph==2.2.0
TgCrypto==1.2.5
tornado==6.4.1
py-tgcalls==2.0.2
rich==13.7.1
pytz==2024.1
psutil
motor
dnspython
GitPython
bs4
Pillow==9.5.0
pymediainfo
requests
unidecode
filetype
attrify
uvloop
pySmartDL
html2text
removebg
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
gpytranslate
httpx[http2]
gtts
youtube-search-python
pymongo
yt-dlp
wget
python-dateutil
emoji
python-dotenv
ntgcalls==1.2.1
numpy==1.26.4
opencv-python-headless==4.10.0.82
pyromod==1.5
pykeyboard
httpx[http2]==0.27.2
git+https://github.com/hakutakaid/pyrogram
openai
pydub
search_engine_parser
geopy
speedtest-cli